- Add default milling value and range - Add `_reset_project_data` method to reset the data on all tabs - Add `query_terminate` method to tab and tab controller to perform action prior to terminating the tab - Confirm closing the tab if stage is not in Loading position - Make resetting the project upon clicking the 'New Project' button

+++ src/odemis/gui/main.py
@@ -332,6 +332,11 @@ class OdemisGUIApp(wx.App):
             dlg.Destroy()  # frame
             return
 
+        # Check if there's any action to do before tab termination
+        # Do not terminate if returned False
+        if not self.tab_controller.query_terminate():
+            return
+
         for p in self.plugins:
             try:
                 p.terminate()
